% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minuteSharp.R
\name{applyFunctionToAllStocks}
\alias{applyFunctionToAllStocks}
\title{apply the same function to all stocks in test.txt}
\usage{
applyFunctionToAllStocks(f, ...)
}
\arguments{
\item{f}{function to apply}

\item{...}{any other parameters}
}
\description{
apply the same function to all stocks in test.txt
}
