% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activePosition.R
\name{calcOpenPositionForStock}
\alias{calcOpenPositionForStock}
\title{calc position stock}
\usage{
calcOpenPositionForStock(dat, symb)
}
\arguments{
\item{dat}{date}

\item{symb}{stock}
}
\description{
calc position stock
}
