% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpe.R
\name{calcSSSharpe}
\alias{calcSSSharpe}
\title{calculate sharpe of a specific stock
calc sharp}
\usage{
calcSSSharpe(symb, dat)
}
\arguments{
\item{symb}{stock}

\item{dat}{date on which to retrieve sharpe}
}
\description{
calculate sharpe of a specific stock
calc sharp
}
