% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpe.R
\name{compareAllSharpYtdDate}
\alias{compareAllSharpYtdDate}
\title{compare all sharpe}
\usage{
compareAllSharpYtdDate(f, ...)
}
\arguments{
\item{f}{the function to apply to all symbols}

\item{...}{any addition params to be passed to the function f}
}
\description{
compare all sharpe
}
