% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekendTask.R
\name{computeWeekPercReturn}
\alias{computeWeekPercReturn}
\title{computing week percentile and week return}
\usage{
computeWeekPercReturn(symb)
}
\arguments{
\item{symb}{stock}
}
\description{
computing week percentile and week return
}
