% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench.R
\name{fromBenchToBenchlist}
\alias{fromBenchToBenchlist}
\title{processing a bench with four columns (output from getbenchmark) to benchlist.txt
this involves copying and possibly not that efficient}
\usage{
fromBenchToBenchlist()
}
\arguments{
\item{d}{a data.table with 4 columns - ticker chnName benchTicker correl}
}
\value{
does not return anything, but outputs a 3-col (ticker indexNameCorrel index) ()
}
\description{
processing a bench with four columns (output from getbenchmark) to benchlist.txt
this involves copying and possibly not that efficient
}
