% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm.R
\name{getClosingPriceBeforeD}
\alias{getClosingPriceBeforeD}
\title{get closing price}
\usage{
getClosingPriceBeforeD(dat, symb)
}
\arguments{
\item{dat}{date}

\item{symb}{stock symbol}
}
\description{
get closing price
}
