% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench.R
\name{getCorrel}
\alias{getCorrel}
\title{get correl between symb and index}
\usage{
getCorrel(symb1, index, env)
}
\arguments{
\item{symb1}{stock symbol sh510050}

\item{index}{index sh000001}

\item{env}{environment in which to retrieve index}
}
\description{
get correl between symb and index
}
