% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench.R
\name{getCorrelGen}
\alias{getCorrelGen}
\title{correl between one stock and all benches}
\usage{
getCorrelGen(symb, env)
}
\arguments{
\item{symb}{stock name}

\item{env}{environment to get variable}
}
\description{
correl between one stock and all benches
}
