% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcure.R
\name{getDataVol}
\alias{getDataVol}
\title{get data including vol traded and amount}
\usage{
getDataVol(symb)
}
\arguments{
\item{symb}{stock name}
}
\description{
get data including vol traded and amount
}
