% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minuteSharp.R
\name{getDayCumuSharpe}
\alias{getDayCumuSharpe}
\title{get cumulative sharpe of one day for one symbol}
\usage{
getDayCumuSharpe(symb, dat)
}
\arguments{
\item{symb}{stock}

\item{dat}{date}
}
\description{
get cumulative sharpe of one day for one symbol
}
