% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minuteSharp.R
\name{getDaySharpe}
\alias{getDaySharpe}
\title{get the sharpe ratio of the day given a stock and a date}
\usage{
getDaySharpe(symb, dat)
}
\arguments{
\item{symb}{stock}

\item{dat}{date}
}
\description{
get the sharpe ratio of the day given a stock and a date
}
