% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minuteSharp.R
\name{getDaySharpeFromEnv}
\alias{getDaySharpeFromEnv}
\title{get day sharpe with data given in an env}
\usage{
getDaySharpeFromEnv(symb, dat, env)
}
\arguments{
\item{symb}{stock}

\item{dat}{date}

\item{env}{environment from which to getdata}
}
\description{
get day sharpe with data given in an env
}
