% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minuteSharp.R
\name{getDaySharpeSinceDate}
\alias{getDaySharpeSinceDate}
\title{get sharpe from date (inclusive)}
\usage{
getDaySharpeSinceDate(symb, dat)
}
\arguments{
\item{symb}{stock}

\item{dat}{date}
}
\description{
get sharpe from date (inclusive)
}
