% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highLowDates.R
\name{getHighLowDates}
\alias{getHighLowDates}
\title{get the high and low dates of a stock}
\usage{
getHighLowDates(symb)
}
\arguments{
\item{symb}{stock symbol}
}
\description{
get the high and low dates of a stock
}
