% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histHighDate.R
\name{getHistHighDate}
\alias{getHistHighDate}
\title{get the hist high date (earliest match, conservative)}
\usage{
getHistHighDate(symb)
}
\description{
get the hist high date (earliest match, conservative)
}
