% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minuteSharp.R
\name{getIndexDayCumuSharpe}
\alias{getIndexDayCumuSharpe}
\title{get minute cumulative sharpe}
\usage{
getIndexDayCumuSharpe(dat)
}
\arguments{
\item{dat}{date}
}
\description{
get minute cumulative sharpe
}
