% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divNew.R
\name{getLastCloseV3}
\alias{getLastCloseV3}
\title{internal method to get close data}
\usage{
getLastCloseV3(symb, dat)
}
\arguments{
\item{symb}{stock}

\item{dat}{date}
}
\description{
internal method to get close data
}
