% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm.R
\name{getMinuteDataForDay}
\alias{getMinuteDataForDay}
\title{get minute data for day}
\usage{
getMinuteDataForDay(dat, symb, pos)
}
\arguments{
\item{dat}{date}

\item{symb}{stock}

\item{pos}{opening pos for that stock}
}
\description{
get minute data for day
}
