% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm.R
\name{getMinuteDataPureForAllDate}
\alias{getMinuteDataPureForAllDate}
\title{get minute data for all date}
\usage{
getMinuteDataPureForAllDate(symb)
}
\arguments{
\item{symb}{stock symbol}
}
\description{
get minute data for all date
}
