% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm.R
\name{getMinuteDataPureWithPrevClose}
\alias{getMinuteDataPureWithPrevClose}
\title{get minute data}
\usage{
getMinuteDataPureWithPrevClose(dat, symb)
}
\arguments{
\item{dat}{date}

\item{symb}{stock symbol}
}
\description{
get minute data
}
