% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm.R
\name{getMinuteDelta}
\alias{getMinuteDelta}
\title{get minute delta}
\usage{
getMinuteDelta(dat, symb, pos)
}
\arguments{
\item{dat}{date}

\item{symb}{stock}

\item{pos}{open position}
}
\description{
get minute delta
}
