% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm.R
\name{getMinuteMtm}
\alias{getMinuteMtm}
\title{get minute mtm}
\usage{
getMinuteMtm(dat, symb, pos)
}
\arguments{
\item{dat}{date}

\item{symb}{stock}

\item{pos}{opening pos for that stock}
}
\description{
get minute mtm
}
