% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openpnl.R
\name{getOpenPos}
\alias{getOpenPos}
\title{get open position with ytd close price
get open pos}
\usage{
getOpenPos(dat)
}
\arguments{
\item{dat}{date}
}
\description{
get open position with ytd close price
get open pos
}
