% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getPercentileCpp}
\alias{getPercentileCpp}
\title{take a data frame and get percentile}
\usage{
getPercentileCpp(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
take a data frame and get percentile
}
