% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minuteSharp.R
\name{getSharpeAllStocksOnDat}
\alias{getSharpeAllStocksOnDat}
\title{get the list of high sharpe stocks on a given date}
\usage{
getSharpeAllStocksOnDat(dat)
}
\arguments{
\item{dat}{date}
}
\description{
get the list of high sharpe stocks on a given date
}
