% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSumChgC}
\alias{getSumChgC}
\title{get sumChg, sumChgSq, n, sr of a change series in list}
\usage{
getSumChgC(x)
}
\arguments{
\item{x}{numvec}
}
\description{
get sumChg, sumChgSq, n, sr of a change series in list
}
