% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minuteSharp.R
\name{getSumSumSq}
\alias{getSumSumSq}
\title{export to file to be processed for wtd sharpe}
\usage{
getSumSumSq(symb, dat)
}
\arguments{
\item{symb}{stock}

\item{dat}{date}
}
\description{
export to file to be processed for wtd sharpe
}
