% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minuteSharp.R
\name{getSumSumSqAll}
\alias{getSumSumSqAll}
\title{get all sum and sum sq for a given date
provide sum and sum sq data for wtd sharpe ( to be used in conjunction with pricemapbar)}
\usage{
getSumSumSqAll(dat)
}
\description{
get all sum and sum sq for a given date
provide sum and sum sq data for wtd sharpe ( to be used in conjunction with pricemapbar)
}
