% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentile.R
\name{getWeekVolPercentile}
\alias{getWeekVolPercentile}
\title{get week trade vol percentile}
\usage{
getWeekVolPercentile(symb)
}
\arguments{
\item{symb}{stock}
}
\description{
get week trade vol percentile
}
