% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minuteSharp.R
\name{getWtdSharpe}
\alias{getWtdSharpe}
\title{the last function was cumulative (showing sharpe for each line)
this method only gets the end of week sharpe}
\usage{
getWtdSharpe(symb, dat)
}
\arguments{
\item{symb}{stock name}

\item{dat}{date}
}
\description{
the last function was cumulative (showing sharpe for each line)
this method only gets the end of week sharpe
}
