% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activePosition.R
\name{loadRecurrentStocks}
\alias{loadRecurrentStocks}
\title{load into main namespace all stocks that exceed an active day threshold of 1}
\usage{
loadRecurrentStocks()
}
\description{
load into main namespace all stocks that exceed an active day threshold of 1
}
